fix(ci): suppress Windows unused warnings and fix clippy lints
Add crate-level cfg_attr to allow unused_imports/variables/dead_code
on Windows (macOS-gated code generates many false positives during
cross-compilation). Fix unnecessary cast in network_monitor_windows.rs
and redundant borrow in url_scheme_windows.rs.

+++ src/gui/network_monitor_windows.rs
@@ -107,8 +107,8 @@ fn read_local_time() -> LocalTime {
         .unwrap_or(0);
     let secs_in_day = (secs % 86400) as u32;
     let hour = (secs_in_day / 3600).clamp(0, 23) as u8;
-    let days_since_epoch = (secs / 86400) as u64;
-    let weekday = (((days_since_epoch + 3) % 7) as u8); // epoch=Thursday, +3→Monday=0
+    let days_since_epoch = secs / 86400;
+    let weekday = ((days_since_epoch + 3) % 7) as u8;
     LocalTime { weekday, hour }
 }
